+++ src/api/auth.py
@@ -4,8 +4,7 @@
     ~~~~~~~~~~~~
 
 """
-from flask import current_app as app, session, redirect
-from src import oauth
+from flask import current_app as app
 from src.api import Blueprint
 from src.common.decorators import authenticate
 from src.common.utils import current_user
@@ -13,38 +12,31 @@ from src.common.utils import current_user
 auth_blueprint = Blueprint("auth", __name__)
 
 
-@auth_blueprint.route("/auth/login")
-def login():
-    # redirect_uri = url_for("auth.getAToken", _external=True)
-    redirect_uri = "http://localhost:8080/popup.html"
-    res = oauth.azure.authorize_redirect(redirect_uri)
-    loc = res.headers.get("Location")
-    return {"loc": loc}, 200
-
-
-@auth_blueprint.route("/auth/getAToken")
-def getAToken():
-    token = oauth.azure.authorize_access_token()
-
-    session["token"] = token
-
-    app.logger.debug(token)
-
-    return redirect(f"http://localhost:8080/popup.html?t={token}")
-
-
-@auth_blueprint.get("/echo")
+@auth_blueprint.get("/auth/me")
 @authenticate
-def echo():
-
-    # token = session.get("token")
-
-    # claims = oauth.azure.parse_id_token(token)
-
-    # roles = claims.get("roles")
+def me():
+    """
+    Returns the currently authenticated user's information
+    ---
+    tags:
+        - auth
+    responses:
+        201:
+            description: ok
+            content:
+                application/json:
+                    schema:
+                        type: object
+                        properties:
+                            roles:
+                                type: array
+                                items:
+                                    type: string
+                                    description: Application roles
+    """
 
     roles = current_user
 
     app.logger.debug(roles)
 
-    return {"roles": roles.get("roles", [])}, 200, [("Access-Control-Allow-Origin", "http://localhost:3000")]
+    return {"roles": roles.get("roles", [])}, 200