+++ src/api/email_verification.py
@@ -9,11 +9,8 @@
         update_verification_status()
 
 """
-from flask import request, redirect, current_app as app
 from src.api import Blueprint
-from werkzeug.exceptions import NotFound, BadRequest
-from src.models.hacker import Hacker
-from src import bcrypt
+from werkzeug.exceptions import Gone
 from src.common.decorators import authenticate, requires_scope
 from src.common.scope import Scope
 
@@ -26,6 +23,9 @@ def check_verification_status(email: str):
     """
     Checks the email verification status
     ---
+    deprecated: true
+    description: >
+        Deprecated, all calls to this endpoint will result in a 410 (Gone)!
     tags:
         - email
     parameters:
@@ -35,6 +35,8 @@ def check_verification_status(email: str):
             type: string
           required: true
     responses:
+        410:
+            description: Gone
         200:
             description: OK
         401:
@@ -43,16 +45,7 @@ def check_verification_status(email: str):
             description: No Hacker exists with that email!
     """
 
-    hacker = Hacker.objects(email=email).only("email_verification").first()
-
-    if not hacker:
-        return NotFound()
-
-    res = {
-        "email_status": hacker.email_verification
-    }
-
-    return res, 200
+    raise Gone("Hacker emails are no longer being verified.")
 
 
 @email_verify_blueprint.get("/email/verify/")
@@ -60,6 +53,9 @@ def update_registration_status():
     """
     Updates the email registration status
     ---
+    deprecated: true
+    description: >
+        Deprecated, all calls to this endpoint will result in a 410 (Gone)!
     tags:
         - email
     parameters:
@@ -69,6 +65,8 @@ def update_registration_status():
             type: string
           required: true
     responses:
+        410:
+            description: Gone
         200:
             description: OK
         404:
@@ -76,27 +74,13 @@ def update_registration_status():
         5XX:
             description: Unexpected error.
     """
-    email_token = request.args.get("token", "")
-    redirect_uri = app.config.get("FRONTEND_URL")
 
-    if not email_token:
-        raise BadRequest("No email token was provided")
-
-    hacker_email = Hacker.decode_email_token(email_token)
-    hacker = Hacker.objects(email=hacker_email).first()
-
-    if not hacker or not hacker.email_token_hash:
-        raise NotFound("Invalid verification token. Please try again.")
-
-    isvalid = bcrypt.check_password_hash(hacker.email_token_hash, email_token)
-    if not isvalid:
-        raise NotFound("Invalid verification token. Please try again.")
-
-    hacker.modify(email_verification=True,
-                  unset__email_token_hash="")
-    hacker.save()
-
-    return redirect(redirect_uri, code=302)
+    raise Gone("Hacker emails are no longer being verified. "
+               "If you have arrived here after clicking the link "
+               "in your email, you don't need to do anything and "
+               "you may safely close this window. "
+               "If you have any questions or are experiencing "
+               "any issues, please contact us at team@knighthacks.org")
 
 
 @email_verify_blueprint.post("/email/verify/<email>/")
@@ -106,6 +90,9 @@ def send_registration_email(email: str):
     """
     Sends a registration email to the hacker.
     ---
+    deprecated: true
+    description: >
+        Deprecated, all calls to this endpoint will result in a 410 (Gone)!
     tags:
         - email
     parameters:
@@ -116,6 +103,8 @@ def send_registration_email(email: str):
             format: email
           required: true
     responses:
+        410:
+            description: Gone
         201:
             description: OK
         404:
@@ -124,17 +113,4 @@ def send_registration_email(email: str):
             description: Unexpected error.
     """
 
-    hacker = Hacker.objects(email=email).first()
-
-    if not hacker:
-        raise NotFound()
-
-    from src.common.mail import send_verification_email
-    send_verification_email(hacker)
-
-    res = {
-        "status": "success",
-        "message": "Verification email successfully sent!"
-    }
-
-    return res, 201
+    raise Gone("Hacker emails are no longer being verified.")