+++ CHANGELOG.md
@@ -18,10 +18,16 @@ Types of Changes:
 
 ### Breaking Changes
 
+-   All authentication and authorization is handled through the Knight Hacks Azure AD tenant. Previous authentication routes will no longer work and will return a 404.
 -   Hacker email verification no longer exists. All requests to `/api/email/*` endpoints will respond with 410 Gone until the next major release.
+
+### Changed
+
+-   Authentication is provided by Azure AD, applications using this API must be granted the `access_as_user` permission.
 
 ### Removed
 
+-   Authentication routes
 -   Hacker email verification.
 
 ## [2.1.9] - 2021-10-25